requirements.txt
absl-py==0.9.0
cachetools==4.0.0
carla==0.9.8
certifi==2018.8.24
chardet==3.0.4
click==7.1.1
configparser==4.0.2
cycler==0.10.0
decorator==4.4.2
dictor==0.1.5
docker-pycreds==0.4.0
elementpath==1.3.3
future==0.18.2
gitdb==4.0.2
GitPython==3.1.0
google-auth==1.11.3
google-auth-oauthlib==0.4.1
gql==0.2.0
graphql-core==1.1
grpcio==1.27.2
idna==2.9
imageio==2.8.0
imgaug==0.4.0
kiwisolver==1.1.0
Markdown==3.2.1
matplotlib==3.0.3
networkx==2.2
numpy==1.18.1
nvidia-ml-py3==7.352.0
oauthlib==3.1.0
opencv-python==4.2.0.32
pandas==0.25.3
pathtools==0.1.2
Pillow==7.0.0
promise==2.3
protobuf==3.11.3
psutil==5.7.0
py-trees==0.8.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydot==1.4.1
pyparsing==2.4.6
python-dateutil==2.8.1
pytorch-lightning==0.7.1
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.3
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-image==0.16.2
scipy==1.4.1
sentry-sdk==0.14.2
Shapely==1.7.0
shortuuid==1.0.1
six==1.14.0
smmap==3.0.1
subprocess32==3.5.4
tensorboard==2.1.1
torch==1.4.0
torchvision==0.5.0
tqdm==4.43.0
urllib3==1.25.8
wandb==0.8.29
watchdog==0.10.2
Werkzeug==1.0.0
xmlschema==1.0.18
